refactor: Landing page on root
